requirements.txt
autopep8
GitPython
pep8
pylint
pylint-django
